Change in filter plugins API: keys are now passed as single char variables and SDL dependency is dropped, in order to have filter actions controlled also from the S-Lang console and to have more abstraction over internally routed opcodes.
This commit breaks backward compatibility with old filters, there might be
some disfunctionality, like Simura effect has no longer control over
colors.
Generally filters will change the useful keytabs in order to stay within
the range of "natural" ASCII chars (no extended modifiers).

This change will be necessary also for the layer's key opcodes.

The filter API will change further (and get stable) once LiViDO filters
will be employed and FreeJ specific filters will be dropped (hope soon!)


git-svn-id: svn://dyne.org/rastasoft/freej/freej@193 383723c8-4afa-0310-b8a8-b1afb83214fc

+++ src/include/jutils.h
@@ -19,7 +19,7 @@
 #ifndef __UTILS_H__
 #define __UTILS_H__
 
-#include <SDL.h>
+#include <inttypes.h>
 #include <console.h>
 
 extern char msg[255];
@@ -39,9 +39,9 @@ void act(char *format, ...);
 void warning(char *format, ...);
 void *jalloc(void *point,size_t size);
 bool jfree(void *point);
-Uint32 fastrand();
-void fastsrand(Uint32 seed);
-unsigned long long int dtime();
+uint32_t fastrand();
+void fastsrand(uint32_t seed);
+double dtime();
 bool set_rtpriority(bool max);
 void jsleep(int sec, long nsec);
 int rtc_open();
